Add parsing for test defs
